refact: add Position/Velocity/Scale component, feat: add display mode; refact: camera & shaders

+++ src/Engine/include/Engine/Window.hpp
@@ -38,7 +38,7 @@ public:
     auto setCursorPosition(glm::dvec2 &&pos) { ::glfwSetCursorPos(m_handle, pos.x, pos.y); }
 
 
-    auto getAspectRatio() const -> float { return static_cast<float>(m_size.x) / static_cast<float>(m_size.y); }
+    auto getAspectRatio() const -> double { return static_cast<double>(m_size.x) / static_cast<double>(m_size.y); }
 
 
     template<typename EventType>
